FIX: corner detection robust to constant images
Some corner detectors contain a division which runs the risk of
dividing by zero when an image is constant (and so has no gradient
in x or y). Adding a small eps (default 1e-6) to these division
instances forces a finite result to be produced without noticeably
affecting the response in other cases.