Radon transform: Include boundary in reconstruction circle.
A test criterion needed to be relaxed slightly to have tests still
passing. This is ok, as the reconstruction circle is now larger, meaning
larger errors should be expected. Moreover, the test in question uses
random data, and changing the seed causes greater changes in accuracy
than the amount the test criterion was relaxed by.