unwrap: Implemented branch cuts.
Tests for small images with wrap around and no masked entries seems
promising. The branch cut core should be ready for masked images and
image borders, but this has not yet been implemented in the setup logic
that defines the problem.

Included with these changes are a queue implementation in C
(the files queue.{c,h}) which is part of the c-algorithms library,
copyright Simon Howard and released under the BSD-like ISC license.