FIX: Fix and improve Poisson random noise generator
The Poissson generator now works.

The improved Poisson generator now infers the bit depth of the image
after conversion to a floating point image, by analyzing the unique
values present and finding the next power of two. This value is then
used to scale the floating point image up, after which Poisson
noise is generated, and then image is then scaled back down.