BUG: wrong search for pivot in `quicker_sort`
This PR corrects an error in the search for the existence of a distinct
pivot for sorting in the `quicker_sort` routines. This could lead to
`quicker_sort` wrongly believing that all items in a subarray are equal and
hence require no sorting.

While this doesn't explain #835 and #902, it may be related.