Warn users of `remove_small_objects` about type
When input type is `int`, we assume that the image is pre-labeled.
However, when only 0 and 1 are present, it may be that the user expects
us to treat it as a binary image. We don't, but now issue a warning
that they might want to use a bool image.

Fixes #1178.