Added new functionality to remove small holes from images.
This is currently working with **kwargs except for inplace
which I cannot get working. It works with arrays of type int
but returns an array of type bool. Possibly in future add
labelling for arrays of type int. A UserWarning is produced
when using arrays of type int which seems to work normally
but the test created for this does not pick up the warning.

Any assistance on these issues would be helpful. I started
this at EuroScipy 2015 and this fixes issue #1642