Remove lena image from test suite.
Results from the morphology module were compared against output from the
corresponding MATLAB functions, as applied to the lena test image.  This image
does not conform to the Debian Free Software Guidelines and had to be removed.
The output is now tested against results generated with the current version of
scikit-image (0.12.0), which is known to be correct due to the comparison
described above.