Update profile.py
Switch and position of src_col and src_row (dst_col and dst_row).

We use img[row, col] to access the pixel intensity of an image.  Corresponded to x-y coordinate, It should be img[y, x], therefore, a x and y in tuple (x,y) should be assigned to col and row respectively, not the other way around.