Added more specific check for image shape in threshold_otsu warning
Currently a warning is emitted from threshold_otsu if image.shape[-1]
is 3 or 4, as this suggests the image may be a RGB or RGBA image. However,
this warning is often raised unnecessarily. For example, the following
shapes of array raise this warning even though they are unlikely to be a RGB
image:

(500, 4)
(4, 3)
(3,)
(4,)

Modifying the code to check whether len(image.shape) > 2 means that none
of these are caught, but all images that may potentially be RGB are still caught.

I'm aware that the logic may be to raise warnings liberally, but it can cause problems,
particularly for people running with 'warnings as errors' - and I think this alteration
still catches all genuine RGB(A) images.