Restoration: Use fast exponent approximation in NL Means
The NL Means algorithm requires very frequent computation of exponents,
to calculate the weights. Using this approximation reduces the overall
calculation time significantly, without a noticeable reduction in accuracy.
NL Means denoising can take a very long of time on large datasets, so this
reduction in calculation time is crucial.

The paper which describes this fast approximation algorithm can be found
here:
https://schraudolph.org/pubs/Schraudolph99.pdf
(Nicol N. Schraudolph, 1999)

Another paper, with a slight variation of it, is here:
http://dx.doi.org/10.1162/089976600300015033
(Gavin C. Cawley, 2000)

Additional improvements and customizations were made by the author of
this patch.

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>