Watershed segmentation: make usable for large arrays (#2967)
* Watershed segmentation: make usable for large arrays

Numbering of indexes is now 64-bit, to allow usage on arrays with more
than 2**31-1 elements. With volumes as large as (for example) 2048**3,
which is not an extraordinary size for 3D volumes anymore, 32-bit index
numbering can not be used.

Additionally, some related quantities are kept 64-bit (and not converted
to 32-bit, as before) to avoid redundant operations. These quantities
are not a part of large arrays, so there is no noticeable effect on
memory usage.

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>