Update fit.py
Changed assignment of origin, direction from self.params to params. Formerly, params argument would have been ignored not matter if it was None or not. Now, if params is None self.params is used as set in line 120 and 121. If params is not None, it will be used if the assert validates.