Add pyproject.toml, now supported by pip
This lets pip figure out that numpy is required to build skimage,
significantly simplifying the checks that need to be done in setup.py

For backward compatibility, we do not update setup.py here.  But we do
clean up AppVeyor and Travis where possible.

Documentation is also updated to simplify installation instructions, and
to include the requirements/* files, instead of copying their
information.