Modernize build process
This seemed like a lot of hard work from @matthew-brett. The last
dependency that used it was pyside (not pyside 2) and that was only used
in in python <=3.4. We now support only 3.5 therefore this isn't
necessary.
Maybe we should point to some of matthew-brett's instructions in case
future maintainers need this? I fear that if we leave the instructions
in but don't really use them, then the instructions kept on scikit-image
will become too old to be useful.

All packages that need to be installed by pip are now in some of the
requirement files. I had to duplicate `numpy` in both build and
defaults. Thought I think this is OK.

Sometimes cython was specified as `0.23` and othertimes as `0.23.4`.
Everything now specifies `0.23.4`. This version is sprinkled in many
locations, so I have a feeling this will be a problem in the future too.

This now checks that we really only require cython and numpy to build
the system.

Previously, this was only applied to defaults.txt