Set `compute` in `apply_parallel` based on array
If the input is a Dask Array, assume that lazy computation was intended.
If the input is anything other than a Dask Array, assume eager
computation was intended. This should provide a bit more sensible
default behavior for `compute` (particularly if users forget to set it).
Though the `compute` option remains to override should it be needed.