denoise_wavelet: warn the user if the wavelet is not orthogonal.
Biorthogonal wavelets transforms color white noise and the filters used do not have unit
norm. This violates assumptions made by the thresholding routines, leading to suboptimal
performance. We should warn the user if such a wavelet is selected.

Add a note on wavelet choise to the denoise_wavelet docstring

Add test for UserWarning on biorthogonal wavelet for denoising.

add rationale for orthogonal wavelets to the denoise_wavelet notes