Update Li thresholding
- images with np.nan and np.inf values are now processed correctly
- `temp` value becoming negative was impossible, and this line was not
  covered, so removed it
- code contained numerous fixes for integer offset errors. Since we
  deal with floating point, these are unnecessary. The code is now
  much simpler
- it is also more correct: the tolerance offset that was applied
  previously in fact was giving an incorrect threshold. I verify this
  with a new private function, _cross_entropy, which estimates the
  cross-entropy between foreground and background as defined in the
  original paper.

The only fly in this sweet ointment is that for *one* of our test
images, coins, the threshold found by the iterative method is one off
(or rather 0.32316 off, but this happens to fall on either side of an
integer) from the optimal threshold. I believe this is an unavoidable
side effect of images being discretized into integer bins.