Reduce default tolerance in threshold_li (#3622)
* Reduce the default tolerance in threshold_li

Fixes #3605

In that issue, I found that a tolerance of `range / 2**15` was
sufficient. However, the only thing that guarantees convergence is the
smallest intensity difference in the image (it's impossible for the
update to change if the threshold doesn't change). I've set this as the
new default, as the only downside is that it might take a long time to
converge, but that seems preferable to finding an incorrect threshold.

* Fix documentation of tolerance in threshold_li