Round instead of ceil output shape when rotating image (#3173)
Round instead of ceil output shape when rotating image

Previously, we ensured that the output contained every subpixel information
from the input when rotating by ceiling the shape. This lead to inconsistent
output when rotating by multiples of 90 degrees. For example:

    >>> transform.rotate(np.zeros((470, 230)), angle=90, resize=True).shape
    (231, 470)

The new behavior correctly produces an output shape of (230, 470).