Fix border removal in peak_local_max with footprints
The current code will erroneously remove the last row of an image when the footprint is size 1 along any dimension.  This is because `remove` will be 1, and (-1 // 2) is -1, and we slice out (-1, None).

In this PR, we update the `remove` to actually reflect the number of pixels we want to trim on each side of the image, and if it is 0, then we continue the loop.

Added tests to put a peak in each pixel in a 3x3 image, and verified that the peaks are correctly detected in all locations.  Prior to the fix, it would fail on any of the last row and last column.