API: Add cross_correlation regularization to register_translation
Because:
- register_translation(A, B) != -register_translation(B, A), but
  it should be!

Shifts are chosen by using the "the first" maximum of the
cross_correlation function, but this means that the shifts are always
biased towards one of the corners of the image. Instead when multiple
shifts have the same cross_correlation, we should choose the smallest
shift.

We can favor smaller shifts by adding a regularizer to the
cross_correlation which favors higher overlap between the two images.

Closes scikit-image/scikit-image#4500