MNT: remove duplicate nogil specification
The functions in _flood_fill_cy were both marked as safe to be called
without the gil at the function level and actually released the gil
with a context manager inside.  Future versions of cython will forbid
this so remove the marker on the function and continue to release the
gil inside.

closes #4545