Alternative update of install page
Install choices are listed according to needs and skill; prerequisites
are given to keep novices from getting in over their head.
If a user wants a pip install, it's made clear what skills
they should master first.

We've been hoping a single doc can do more than it's possible for
a single doc to do. It can't give clear general instructions if
it's also quickstarting newbies and leading bewildered users
out of the weeds. The ambitions are totally worthy but nobody
will win.