fix Azure Pipelines, pytest config, and trove classifiers for Python 3.8
Azure Pipelines did not appear to be running the correct Python versions
because they were running on a Windows image and the UsePythonVersion
task did not set the path correctly for running bash scripts. This
change uses the variable produced by UsePythonVersion to specify the
full path to Python.

This change also updates the trove classifiers to specify Python 3.8
which is already supported and tested, and updates the pytest config to
not try running the test setup as a test itself.