Removed the use of `np.random.choice([True, False], ...)`.
This significantly improves the performance of the salt and pepper
random noise, both by reducing the time taken to generate the random
noise, as well as reducing the peak memory usage.

The new method also actually uses the `seed` provided to the function,
unlike the previous method, so that noise can now be added
deterministically.