DOC: Attempt at autoformatting docstrings in morphology/*
Numpydoc is specific about the meaning of spaces arround : and a few
other things, while it may look almost identical in sphinx depending on
themes it is semantically different.

I'm attempting to write an docstring autoreformatter thetry to fix
common mistakes, and fix a couple of other things like typo in parameter
names, remove parameters that are still documented, and try to get a
uniform style.

You'll see in the diff
 - add / removal of spaces
 - add / removal of blank lines
 - indent 4 spaces always.
 - Capitalisation of section names
 - removal of `border_value` parameter which is indeed not a function
    argument.

In morphology/* this hasn't found any other issue it can fix, and the
only other non-automatically fixable issues are missing documentation
for `dtype` parameters.