Many test modules had setup and teardown functions defined, but pytest was not using these.
Move the calls to the setup_test and teardown_test utilities into the
top level conftest.py file instead, so they get used for all tests.

This resulted in a number of new failures due to uncaught warnings that
had to be fixed.

Some uncaught warnings are still being skipped in the io module, but
should be looked into further.