Make SLIC superpixel output robust to image scale
closes gh-5512
Currently img_as_float rescales the range of integer inputs, but leaves floating point range unnormalized.
The compactness parameter needs to be tuned based on image scale, so it is based to just always rescale
to range [0, 1] internally so the default is a reasonable starting point regardless of the input image
scale.