Fix #4766
The offsets attribute is needed to interpret the traceback returned
by MCP.find_costs(). The attribute was not declared as public in
the cython code however so it could not be accessed. The attribute
is also documented as being part of the API. This commit makes it
public and adds test coverage for it.