Allow the weighting of point pairs in geometric transforms
If locations of the points pairs used for estimating geometric
transform matrices have come from some sort of fitting or estimation
process themselves, then there may be uncertainties or confidence
values associated with them. By limiting the specification of the
weights to only one value per pair of points, the implementation is
simplified.