MAINT: remove use of distutils.version
using packaging's version.parse instead

packaging is a very lightweight package that is quick to import (unlike
pkg_resources). It is already a dependency of matplotlib, but I don't think it
is a dependency of any of the packages in requirements/default.txt, so I added
it there to make sure it will get installed.