remove all use of the deprecated distutils package (#6044)
* prefer import CompileError, LinkError from setuptools

Now that distutils is deprecated and will be removed in Python 3.12, we
should import these from setuptools instead. They were added to the
public setuptools API in version 59.0.

* MAINT: remove use of distutils.version

using packaging's version.parse instead

packaging is a very lightweight package that is quick to import (unlike
pkg_resources). It is already a dependency of matplotlib, but I don't think it
is a dependency of any of the packages in requirements/default.txt, so I added
it there to make sure it will get installed.

* also require packaging at build time

* add packaging to pyproject.toml as well

(cherry picked from commit 0f67fa74f9b7152aa109e9b8df9a63ca6b3af68f)