Remove use of channel_as_last_axis from apply_parallel (#5585)
* Remove use of channel_as_last_axis from apply_parallel

This commit improves how the `channel_axis` parameter was implemented for
`apply_parallel`. First, it removes the need for the `channel_as_last_axis`
decorator. That decorator resorts the image dimensions with moveaxis, but
would not make corresponding shifts in the elements of `chunks`, `depth`,
etc. if they are specified. This can make it quite confusing to use when
channel_axis != -1.

A second improvement is allowing `depth` and `chunks` to optionally be a
sequence corresponding to only the non-channel dimensions. A depth of 0 will
be automatically inserted for the channels dimensions. Similarly a chunk
size of `array.shape[channel_axis]` will be used on the channel_axis
(e.g. no chunking along channels).

TST: exand parameterized cases substantially for apply_parallel

test over additional chunks and depth inputs as well as additional
channel_axis settings.

* Apply suggestions from code review

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>