ENH: add true Butterworth variant
The current Butterworth filter is actually the square of a Butterworth filter.
Add a square_butterworth keyword argument defaulting to True for backwards compatibility.
This argument can be set to False to allow use of the standard Butterworth.
Some textbooks use the squared form, but this is not how a Butterworth is
defined in signal processing.

Added test cases verifying the expected amplitudes at the cutoff in each case.

DOC: Add an example using the Butterworth filter to the gallery

DOC: clarify definition of cutoff frequency ratio in the docstring

ENH: raise a ValueError when cutoff_frequency_ratio is out of range

DOC: Fix sign in the denominator of the equations in the docstring and
specify these explicitly for the low-pass vs. high-pass cases

ENH: add npad kwarg to make image padding more convenient