Handle memory layout more expicitely
`out` must be C-contiguous which allows for efficient indexing in the
underlying Cython code. This might raise an error if out is supplied by
the user so document it with a test.

It seems that special logic for bool-type arrays is no longer necessary
as the appropriate PR [1] was merged.

[1] https://github.com/cython/cython/pull/2676