Use mask during rescaling in slic
Previously, the mask was ignored when rescaling the image to make choice
of compactness insensitive to the image values. So contrary to the
docsting, NaN's (and infinite values) couldn't actually be masked.

Therefore, we use a masked array to compute the max and min value for
rescaling. Additionally these intermediate results are checked for
NaN and infinite which would give wrong rescaling results.

This solution needs to invert the mask, because NumPy's masked array
and this function interpret True & False the opposite way.