Raise ValueError for 2D multichannel input to slic
The way I understand the docstring, an image must be at least 2D.
channel_axis=-1 (the default) indicates that the last dimension is a
multichannel / color channel which would make the image 1D. Therefore,
this parameter constellation should raise a ValueError.

This change adds a guard for this case, an appropriate test and fixes
several unit tests which relied on the fact that the implementation
ignored channel_axis for 2D images.