Use mask during rescaling in segmentation.slic and improve handling of error cases (#6525)
* Use mask during rescaling in slic

Previously, the mask was ignored when rescaling the image to make choice
of compactness insensitive to the image values. So contrary to the
docsting, NaN's (and infinite values) couldn't actually be masked.

Therefore, we use a masked array to compute the max and min value for
rescaling. Additionally these intermediate results are checked for
NaN and infinite which would give wrong rescaling results.

This solution needs to invert the mask, because NumPy's masked array
and this function interpret True & False the opposite way.

* Check that slic ignores offset during rescaling

* Raise ValueError for 2D multichannel input to slic

The way I understand the docstring, an image must be at least 2D.
channel_axis=-1 (the default) indicates that the last dimension is a
multichannel / color channel which would make the image 1D. Therefore,
this parameter constellation should raise a ValueError.

This change adds a guard for this case, an appropriate test and fixes
several unit tests which relied on the fact that the implementation
ignored channel_axis for 2D images.

* Use boolean indexing instead of mask

* Check NaN and infinity before calculation

and prevent exceptions like
RuntimeWarning: invalid value encountered in divide
if imin == imax.

* Add channel kwarg for slic's grayscale benchmark

This argument is necessary because of the previously introduced check
that expects channel_axis=None if a grayscale image is passed to slic.

* Reword error message for 2D and multichannel

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>