Start deprecation API inconsistency
in phase_cross_correlation. Depending on return_error and if masked
cross-correlation is used, the function returns either shifts or a tuple
of shifts, error and phasediff. This API inconsistency shall be
deprecated in favor of always returning shifts, error and phasediff.

Deprecation plan:
- This release (0.20?): raise a warning that next release will start
  returning error and phasediff (as NaN if mask is supplied). The
  warning goes away with return_error="always", which also changes to
  the new behavior of always returning error and phasediff.
- Next release (0.21 or 2.0?): deprecate return_error, and make
  returning 3 outputs the default.

We warn only were a single output is returned as opposed to doing a
general warning for return_error != "always". This saves users who are
already using the function with 3 outputs from having to update their
use of the function.