Continue work on release notes
I'm confused with what to do abou the Milestone 1.0. Should these be
included in 0.20 now. Collect them for now until that's confirmed.

Furthermore, collect the backported PRs. We can delete them later but
for now it's useful to keep them for review and in case we want to go
back. Note that quite a lot are missing because I didn't start this
earlier.