Include dev forum mail address in documentation
Mailing skimage@discuss.scientific-python.org is a way to create a post
on our forum without having an account [1].

I added skimage@discuss.scientific-python.org wherever we mention
"getting in touch" or "reaching out" on our developer forum. This
should lower the barrier a bit for new users how don't have an account
yet. Not sure if we have to worry about spam or how we would address
that? Setting the approval level limit to 1 instead of 0 might to be a
way to mitigate this if it becomes a problem [2]. Or we will have to
turn off anonymous senders again.

[1] https://meta.discourse.org/t/configure-incoming-email-to-create-new-topics-or-group-messages/62977
[2] https://meta.discourse.org/t/allowing-topic-creation-through-direct-delivery-email-in-makes-my-forum-vulnerable-to-spammers/118421