Reorder and -structure release notes
This introduces additional subsections as a suggestion. Hopefully this
makes it easier for readers to focus on the stuff they are actually
interested in without having to read through a list of minor fixes and
updates.

Were it made sense I also consolidated related bullet points. E.g. if
one PR added something new and a following one added some further fixes
or improvements. In that case both PR numbers are appended in brackets.