Replace skimage_deprecation with FutureWarning
This was only used in the decorator `deprecated` while everywhere else
FutureWarning is used. So let's be consistent about this.

Also don't modify the warning filter when calling the wrapped function.
That this happens is very implicit and hidden from the user. I'd argue
that it's the users responsibility to configure his warning filter while
it's our responsibility to chose good and expected defaults.