Update decorator for deprecated functions
Previously, the warning did not point to where the deprecated function
was called(user code) but at the deprecated function's definition
(library code)! Add a test for this.

I also propose an updated signature of the decorator, that makes sure
that the deprecated version is specified (forgotten in previous uses
of this decorator). I replaced `alt_func` with a more general `hint`
parameter as there is not always a clear alternative function. In that
case it's possible to provide a more flexible guiding message to the
user.

Also do away with the behavior parameter. Raising an exception defeats
the purpose of a deprecation decorator as it immediately breaks user
code. It impossible to use the deprecated function, in which case it
should just be removed.