Deprecate automatic channel detection in gaussian
Automatic detection of the color channel based on the old deprecated
`multichannel=None` was broken in version 0.19. This commit recovers
the old behavior for 0.20 using a proxy object `ChannelAxisNotSet`.
The behavior is still deprecated though and should be removed completely
in 0.21.

Note that the warning is only raised if the automatic behavior deviates
from the behavior that will be the new default in 0.21, which is
`channel_axis=None`. This way we only notify users that were affected
by the broken behavior in 0.19 and now deprecated behavior in 0.20.