Use explicit image tag to fix CircleCI job
Also venv doesn't seem to be available by default on the current
ubuntu-2204 machine. We could apt-get it but that resulted in a blocking
prompt to restart some system services. I didn't try to resolve it
because I think that we don't really need a venv here.

See https://circleci.com/developer/machine/image/ubuntu-2204