Refactor _invariant_denoise to denoise_invariant
This private function is used in the plot_j_invariant_tutorial gallery
example. It's useful so make it public and more consistent with the
other denoise_* functions.

The "Notes" and "Examples" section are adapted from the same gallery
example.

I opted to keep the function in this module rather than moving it to
_denoise.py to keep git blame intact.