Fix doc build on CircleCI and add ccache (#6646)
* Use explicit image tag to fix CircleCI job

Also venv doesn't seem to be available by default on the current
ubuntu-2204 machine. We could apt-get it but that resulted in a blocking
prompt to restart some system services. I didn't try to resolve it
because I think that we don't really need a venv here.

See https://circleci.com/developer/machine/image/ubuntu-2204

* Add build caching with ccache

This is adapted from SciPy's config [1]. Ccache seems to be picked up
by meson but it's not really visible from the output if caching is
successfully used. However, the CI step is shorter (3 min compared to
previous 5 min).

[1] https://github.com/scipy/scipy/blob/121da759effddeda6986ef09f0f112f099815509/.circleci/config.yml

* Print ccache performance