Make 3D test of hessian_matrix more specific
Explanation: Each array in Hs is the result of

    H_elems = [np.gradient(gradients[ax0], axis=ax1)
               for ax0, ax1 in combinations_with_replacement(axes, 2)]

Because reversing axes (due to the bug fix) doesn't affect the result
of `H_elems[2]`. For index 2, `combinations_with_replacement(axes, 2)`
returns either `(0, 2)` or the reversed `(2, 0)`. Both result in the
same gradient.

So to make the test notice this change, we test another one of H_elems
arrays whose result is not the same if axes is reversed.