Fix RAG.merge_nodes for in-place merging
When merging in-place, `new` is assigned a new node ID that doesn't yet
have any edges to pre-existing nodes. So the `wheight_func` can only
ever identify the existing edge neighbor-src; the edge neighbor-new
doesn't exist yet.

That it worked for the `in_place=True` was because `new` was set to
`dst`.