Restructure documentation toctree
The pydata-sphinx-theme uses a layout where the first toc level is used
in the header bar. This limits the number of top-level sections that are
practical.

This reorganization is a first attempt to that sorts everything into
the top-level sections "User guides", "Examples", "API reference",
"Release notes" and "Project". This losely follows what other projects
such as SciPy, NumPy, networkx do.

Also introduces sphinx-design as a new dependency which provides the
grid directive that is used on the documentations landing page.