Fix struct warning in detect_multi_scale (#6728)
Fixes the warnings:

warning: skimage/feature/_cascade.pyx:753:49: Not all members given for struct 'Detection'
warning: skimage/feature/_cascade.pyx:753:49: Not all members given for struct 'Detection'

The memory for the Detection struct was already allocated on the stack
of this function, so we can just directly assign to its members. I also
tried the syntax

    new_detection = Detection(
        r=current_row
        c=current_col
        width=current_width
        height=current_height
    )

with and without parameter names. However, that leads to the following
compilation error:

/home/lg/Res/scikit-image/skimage/feature/_cascade.pyx:753:50: Operation not allowed without gil