BUG: Fix heap growing memory error and silence compiler warning (#6727)
* BUG: Fix heap growing memory error and silence compiler warning

I am not 100% sure how certain the fix is here.  The code seemed always
fine (although I am not quite sure that pointer arithmetic on dead pointers
is technically valid).

The current logic seemed to be broken for realloc failures, so I fixed that
which is the larger diff...

(I have not added tests since trigger memory error in the right place is
often hard, but could have a try...)

xref gh-6725

* BUG: Add `with gil:` to heap realloc memoryerror raising

* BUG: allow heappush to propagate errors explicitly

* BUG: Also propagate MemoryError correctly from queue_grow_buffer

Otherwise, cython will end up printing the error because it can't
raise it correctly.