MAINT: Use offsets rather than pointer during heap realloc
This converts pointers to offsets (storing them in-place) during
heap reallocation and then recalculates the new pointers after the
realloc.

That may well be the only correct way to do the realloc here based
on what the C standard guarantees.  In practice, I doubt that sklearn
can target hardware where this particular use-case matters (I am not
even sure such hardware exists, the `uintptr_t` cast itself would have
to rely on the pointers not being dangling for this to be unsafe, but
I guess such hardware may be plausible...).