Fix threshold_li(): prevent log(0) on single-value background.
When background mode has only a single value (normalized to 0), its
logarithm is not defined and produces a warning or exception and the
resulting threshold was the same as the background.
Instead we can stop the iteration and take the last threshold which is
between the two modes.

A few examples of bi-modal arrays with single-value background are
provided to the existing unit test (both for byte or float images).
Ignoring the numpy warning is no longer needed.