Always include minigallery detective
At the time of generating the API source files with
tools/build_mod_ref.py the gallery has not yet been build so we can't
know in advance which functions have a gallery example to link back to.
An option is to just always include the minigallery directive which
makes sure that existing ones are included and does nothing if none
are found.

Though it would be nice to to be able include a heading with
"rubric:: Gallery examples" if backrefs exist. It renders as if it
were part of the docstring. We might be able to solve that with
templating...