Use labels as input and map priority to index
This overhauls the API of remove_near_objects. Instead of labeling the
input itself, the function already expects a labeled image. Sorting an
index into this label_image by label ID first, and by priority second,
makes it easy to find other indices belonging to the same object because
they are contiguous in the resulting index. This makes it possible to
entirely omit the parameters connectivity and footprint (for now).

`priority` now is expected to assign a priority to each object and no
longer gives a priority for each pixel.

Test and benchmarks were updated accordingly.