Add merge_group trigger to workflow config
We plan to enable merge queues for our pull requests. With this GitHub
feature enabled, pull requests can be added to a merge queue. GitHub
then creates a temporary branch between pull request and target branch
(main) and runs the checks. This might offer some convenience for the
“I’ll merge once the CI is green” cases.

GitHub sends a different event when a pull request is added to a merge
queue, so we need to update our CI configuration to trigger on
"merge_group" events [1].

[1] https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue