Avoid installation of rtoml via conda (#6792)
This is a short-term fix following #6781. There are only two dependencies listed in requirements/developer.txt, i.e., pre-commit and rtoml, which is only available from PyPI .

In practice, removing this dependency for the conda user won't be blocking, since it's only used for pyproject.toml, and the import failure is already handled.