Relax reproduce section in bug issue template
The previous formatting forced the "Way to reproduce:" and "Traceback
or output:" sections to use simple text fields without rich Markdown
formatting. I noticed that users tend to just use the description field
instead.

I think we should allow contributors some degree of freedom in
describing how to reproduce the bug. Sometimes it might not be enough to
post a simple code snippet and more complex behavior needs to be
described. Or the  contributor wants to compare correct and faulty
behavior next to each other.

This also makes it unnecessary to have a separate field for the output.