Apply pre-commit only to diff in CI
This allows to set stricter checks without having to apply it on the
whole code base and update over time. We can do so by taking advantage
of the --from-ref and --to-ref options in pre-commit run, to select the
diff that the checks are performed on.

GITHUB_BASE_REF and GITHUB_HEAD_REF only exist for pull requests [1].
For workflows triggered by the merge queue we can use GITHUB_REF_NAME to
find the base by stripping the prefix "gh-readonly-queue/" [2].
Otherwise we fall back to checking the last commit.

[1] https://docs.github.com/en/actions/learn-github-actions/variables#default-environment-variables
[2] https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue#about-merge-queues