Expose `color.get_xyz_coords` in public API (#6696)
* Expose color.get_xyz_coords in public API

For now it's mentioned once in the docstring of `skimage.color.lab2rgb`
but it's probably intended to be part of our public API. So it should
be part of our public API documentation too!

See also gh-6688 which directly references this function.

* Refactor get_xyz_coords to xyz_tristimulus_values

This seems to be the term actually used by the reference in the
function's docstring. Add a - in my opinion - more useful reference
as well.

* Revert unintended change to references

These links were updated in a previous PR (to "CIELAB_color_space"). No
idea, how this change was reverted. Maybe during resolving of the merge
conflicts.

* Add notes to xyz_tristimulus_values

This explanation is adapted from the description of the "_illuminants"
dictionary [1]. The formulars were modified to take the scaling of Y to
1 into account.

I'm not sure, how to explain the values of the D65 / 2° combo or what
to do about it.

[1] https://github.com/scikit-image/scikit-image/blob/8495b071ff81f3f1c5c4d8a44cb0387a81092066/skimage/color/colorconv.py#L450-L500

* Temporarily add deprecated get_xyz_color

* Fix version number in TODO.txt

Likely a typo. See 7b448105017e5aed286389661c2aa4397cb37c88
which uses removed_version=0.23 [1].

[1] https://github.com/scikit-image/scikit-image/commit/7b448105017e5aed286389661c2aa4397cb37c88

* Fix doctest in xyz_tristimulus_values