Showcase pydata-sphinx-theme (#6714)
* Restructure documentation toctree

The pydata-sphinx-theme uses a layout where the first toc level is used
in the header bar. This limits the number of top-level sections that are
practical.

This reorganization is a first attempt to that sorts everything into
the top-level sections "User guides", "Examples", "API reference",
"Release notes" and "Project". This losely follows what other projects
such as SciPy, NumPy, networkx do.

Also introduces sphinx-design as a new dependency which provides the
grid directive that is used on the documentations landing page.

* Switch to pydata_sphinx_theme

Delete the old scikit-image specific theme and templates and switch to a
minimal configuration of the pydata_sphinx_theme.

* Tweak gallery example intro text

* Remove release_notes job in doc/Makefile

For now and for demo purposes we curate an index file manually that
includes all past releases. In the future we should probably clean
up the confusing section levels.

* Clean up generated API page

The TOC in the left sidebar of the new pydata_sphinx_theme was very
cluttered with unnecessary entries. This removes them an also integrates
the backrefs to gallery examples more nicely; they look identical to
other docstring sections now.

* Add ~= constraint to build_versions.py

* WIP

* Always include minigallery detective

At the time of generating the API source files with
tools/build_mod_ref.py the gallery has not yet been build so we can't
know in advance which functions have a gallery example to link back to.
An option is to just always include the minigallery directive which
makes sure that existing ones are included and does nothing if none
are found.

Though it would be nice to to be able include a heading with
"rubric:: Gallery examples" if backrefs exist. It renders as if it
were part of the docstring. We might be able to solve that with
templating...

* Address linting errors

found by ruff.

* Fix reference to user guides

* Update the start page

Try to retain most of the useful links of the old version and make
the descriptions in the quick link section a bit more concise.

* Include the governance SKIP twice

Include the SKIP directly in the Community toctree to highlight a bit
more how our project is governed.

* Move version switcher behind logo

* Use GitHubGist link to version_switcher.json

_static/version_switcher.json only works on the root page of the
generated HTML. Try it this way to also checkout how the switcher deals
with the updated site structure.

* Rename to singular "User guide"

* Number documents in User guide

* Override default PyData Pygments style

Stéfan didn't like the default style "tango" which PyData uses [1].
Instead, I opted for the styles that meet the WCAG 2.1 AA contrast
minimum [2]. Unfortunately this requires additional styling to ensure
links in code cells stay black as intended by the style.

[1] https://pydata-sphinx-theme.readthedocs.io/en/latest/user_guide/styling.html#configure-pygments-theme
[2] https://pygments.org/styles/

* Replace logo with icon and text

This way we can use the dark theme as well.

* Add & set 0.20 as stable to version switcher

* Update generated pyproject.toml

* Split project into About and Developer guide

* Allow 6 header links before dropdown

The default "header_links_before_dropdown" is 5. As the theme uses
6 total, the drop-down only has a single item. Instead we can just
replace the dropdown with the original item.

* Rename developer guide to development

This follows SciPy and the more generic term is more fitting as this
section not only contains guides but also SKIPs.

* Unify release notes titles

* Include SKIP template in visible TOC

Even though it's hidden, it would still show up in the left sidebar. We
could mark the template as :orphan: and not include it in the TOC,
though people would have to remember to remove that field when copying
the template. This seems easier.

* Add section descriptions

* Use less vertical space in footer

* Hide toctree on start page

* Automatically include release notes

* Fix typos and tweak short section descriptions

* Apply rewording suggestions

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>

* Allow sphinx>=5 until sphinx_design supports 6.0

The sphinx dependency was updated in
7b3ed09281f3ae3de203f7b42f4c20006a0c0597.

However, for this feature branch introduces sphinx_design 0.3.0 which
doesn't yet support sphinx>=6 [2]. Note, the current 'main' branch of
sphinx_design already supports sphinx>=6, so the next release should
allow us to upgrade.

[1] https://github.com/scikit-image/scikit-image/commit/7b3ed09281f3ae3de203f7b42f4c20006a0c0597
[2] https://github.com/executablebooks/sphinx-design/blob/42077cf08756019db3936fbd9a66d68f6ce6d23d/pyproject.toml#L30

---------

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>